style: kill some warnings

+++ Project2_Simple_Kernel/test/test_project2/add.c
@@ -35,8 +35,8 @@ int main(void)
         data[i]=i;
     }
 
-    sys_thread_create((uint64_t)add1, &finished1, &ans1, 0, 0);
-    sys_thread_create((uint64_t)add2, &finished2, &ans2, 0, 0);
+    sys_thread_create((uint64_t)add1, (long)&finished1, (long)&ans1, 0L, 0L);
+    sys_thread_create((uint64_t)add2, (long)&finished2, (long)&ans2, 0L, 0L);
 
     while(!finished1||!finished2);
     sys_move_cursor(0, print_location);
